requirements.txt
Flask==3.0.0
pandas==2.3.2
numpy>=2.0.0
scikit-learn>=1.5.0
pyecharts==2.0.3
requests==2.32.3
beautifulsoup4==4.12.3
joblib>=1.4.0
python-dotenv==1.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.3
Werkzeug>=3.0.0
pymysql>=1.1.0
